+++ source/Relativity.Testing.Framework.Api/Services/IGroupService.cs
@@ -15,10 +15,10 @@ namespace Relativity.Testing.Framework.Api.Services
 	public interface IGroupService
 	{
 		/// <summary>
-		/// Creates the specified <see cref="Group"/>.
+		/// Creates the specified [Group](https://relativitydev.github.io/relativity.testing.framework/api/Relativity.Testing.Framework.Models.Group.html).
 		/// </summary>
-		/// <param name="entity">The <see cref="Group"/> entity to create.</param>
-		/// <returns>The created <see cref="Group"/> entity.</returns>
+		/// <param name="entity">The [Group](https://relativitydev.github.io/relativity.testing.framework/api/Relativity.Testing.Framework.Models.Group.html) entity to create.</param>
+		/// <returns>The created [Group](https://relativitydev.github.io/relativity.testing.framework/api/Relativity.Testing.Framework.Models.Group.html) entity.</returns>
 		/// <example>
 		/// <code>
 		/// var client = relativityFacade.Resolve&lt;IClientService&gt;().Get(clientArtifactId);
@@ -35,9 +35,9 @@ namespace Relativity.Testing.Framework.Api.Services
 		/// <summary>
 		/// Requires the specified group.
 		/// <list type="number">
-		/// <item><description>If <see cref="Artifact.ArtifactID"/> property of <paramref name="entity"/> has positive value, gets entity by ID and updates it.</description></item>
-		/// <item><description>If <see cref="NamedArtifact.Name"/> property of <paramref name="entity"/> have a value, gets entity by name and updates it if it exists.</description></item>
-		/// <item><description>Otherwise creates a new entity using <see cref="ICreateWorkspaceEntityStrategy{T}"/>.</description></item>
+		/// <item><description>If [ArtifactID](https://relativitydev.github.io/relativity.testing.framework/api/Relativity.Testing.Framework.Models.Artifact.html#Relativity_Testing_Framework_Models_Artifact_ArtifactID) property of <paramref name="entity"/> has positive value, gets entity by ID and updates it.</description></item>
+		/// <item><description>If [Name](https://relativitydev.github.io/relativity.testing.framework/api/Relativity.Testing.Framework.Models.NamedArtifact.html#Relativity_Testing_Framework_Models_NamedArtifact_Name) property of <paramref name="entity"/> have a value, gets entity by name and updates it if it exists.</description></item>
+		/// <item><description>Otherwise creates a new entity using <see cref="ICreateWorkspaceEntityStrategy&lt;T&gt;"/>.</description></item>
 		/// </list>
 		/// </summary>
 		/// <param name="entity">The entity to require.</param>
@@ -86,19 +86,21 @@ namespace Relativity.Testing.Framework.Api.Services
 		/// Gets the group by the specified ID.
 		/// </summary>
 		/// <param name="id">The artifact ID of the group.</param>
-		/// <returns>The <see cref="Group"/> entity or <see langword="null"/>.</returns>
+		/// <param name="includeMetadata">Indicates wheter to include group Meta property. Default is false.</param>
+		/// <param name="includeActions">Indicates wheter to include group Actions property. Default is false.</param>
+		/// <returns>The [Group](https://relativitydev.github.io/relativity.testing.framework/api/Relativity.Testing.Framework.Models.Group.html) entity or <see langword="null"/>.</returns>
 		/// <example>
 		/// <code>
 		/// var entity = _groupService.Get(groupArtifactId);
 		/// </code>
 		/// </example>
-		Group Get(int id);
+		Group Get(int id, bool includeMetadata = false, bool includeActions = false);
 
 		/// <summary>
 		/// Gets the group by the specified group name.
 		/// </summary>
 		/// <param name="name">The name of the group.</param>
-		/// <returns>The <see cref="Group"/> entity or <see langword="null"/>.</returns>
+		/// <returns>The [Group](https://relativitydev.github.io/relativity.testing.framework/api/Relativity.Testing.Framework.Models.Group.html) entity or <see langword="null"/>.</returns>
 		/// <example>
 		/// <code>
 		/// var name = "Some Existing Group Name";
@@ -111,7 +113,7 @@ namespace Relativity.Testing.Framework.Api.Services
 		/// Gets all groups by the specified names.
 		/// </summary>
 		/// <param name="names">The collection of group names.</param>
-		/// <returns>The collection of <see cref="Group"/> entities.</returns>
+		/// <returns>The collection of [Group](https://relativitydev.github.io/relativity.testing.framework/api/Relativity.Testing.Framework.Models.Group.html) entities.</returns>
 		/// <example>
 		/// <code>
 		/// var names = new List&lt;string&gt;{"Some Existing Group Name", "Other Existing Group Name"};
@@ -124,6 +126,7 @@ namespace Relativity.Testing.Framework.Api.Services
 		/// Updates the specified group.
 		/// </summary>
 		/// <param name="entity">The entity to update.</param>
+		/// <returns>Updated <see cref="Group"/>.</returns>
 		/// <example>
 		/// <code>
 		/// var groupId = 1;
@@ -132,9 +135,9 @@ namespace Relativity.Testing.Framework.Api.Services
 		/// 	Name = "Some Existing Group Name",
 		/// 	Keywords = "Test Edited Keywords"
 		/// }
-		///  _groupService.Update(entity);
+		/// Group updatedGroup = _groupService.Update(entity);
 		/// </code>
 		/// </example>
-		void Update(Group entity);
+		Group Update(Group entity);
 	}
 }