+++ source/Relativity.Testing.Framework.Api/Services/Imaging/IImagingDocumentService.cs
@@ -9,7 +9,7 @@ namespace Relativity.Testing.Framework.Api.Services
 		/// </summary>
 		/// <param name="workspaceId">The workspace ID.</param>
 		/// <param name="documentArtifactId">The Document ArtifactID.</param>
-		/// <returns>Return a <see cref="DocumentStatus"/> of imaging job for a document.</returns>
+		/// <returns>Return a [DocumentStatus](https://relativitydev.github.io/relativity.testing.framework/api/Relativity.Testing.Framework.Models.DocumentStatus.html) of imaging job for a document.</returns>
 		DocumentStatus GetStatus(int workspaceId, int documentArtifactId);
 	}
 }