+++ source/Relativity.Testing.Framework.Api/Strategies/Permissions/DTO/PermissionGroupDTO.cs
@@ -1,0 +1,14 @@
+﻿using Relativity.Testing.Framework.Models;
+
+namespace Relativity.Testing.Framework.Api.Strategies
+{
+	internal class PermissionGroupDTO
+	{
+		public PermissionGroupDTO(int groupId)
+		{
+			Group = new Artifact(groupId);
+		}
+
+		public Artifact Group { get; set; }
+	}
+}