+++ source/Relativity.Testing.Framework.Api/Strategies/Scripts/IScriptReadRunJobStrategy.cs
@@ -13,7 +13,7 @@ namespace Relativity.Testing.Framework.Api.Strategies
 		/// </summary>
 		/// <param name="workspaceId">The ArtifactID of the workspace where the script exists and was executed.</param>
 		/// <param name="runJobId">The ID of the script run job.</param>
-		/// <returns>A <see cref="RunJob"/> describing the script run job.</returns>
+		/// <returns>A [RunJob](https://relativitydev.github.io/relativity.testing.framework/api/Relativity.Testing.Framework.Models.RunJob.html) describing the script run job.</returns>
 		RunJob ReadRunJob(int workspaceId, Guid runJobId);
 	}
 }